controller/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
googleapis-common-protos==1.62.0
python-dotenv==1.0.0
dependency-injector==4.41.0